+++ src/main/java/com/brand/blockus/datagen/models/BlockusModels.java
@@ -15,9 +15,11 @@ public class BlockusModels {
     public static final Model TEMPLATE_POST_CONNECT;
     public static final Model TEMPLATE_POST_CONNECT_TOP;
     public static final Model CHAIN_CONNECT;
-    public static final Model TEMPLATE_SMALL_HEDGE_END;
+    public static final Model TEMPLATE_SMALL_HEDGE_NOSIDE;
+    public static final Model TEMPLATE_SMALL_HEDGE_NOSIDE_ALT;
+    public static final Model TEMPLATE_SMALL_HEDGE_POST;
     public static final Model TEMPLATE_SMALL_HEDGE_SIDE;
-    public static final Model TEMPLATE_SMALL_HEDGE_SIDE_TALL;
+    public static final Model TEMPLATE_SMALL_HEDGE_SIDE_ALT;
     public static final Model TEMPLATE_SMALL_HEDGE_INVENTORY;
     public static final Model CRATE_TEMPLATE;
     public static final Model NEON_BLOCK_TEMPLATE;
@@ -44,9 +46,11 @@ public class BlockusModels {
         TEMPLATE_POST_CONNECT = createModel("template_post_connect", "_connect", TextureKey.END, TextureKey.SIDE);
         TEMPLATE_POST_CONNECT_TOP = createModel("template_post_connect_top", "_connect_top", TextureKey.END, TextureKey.SIDE);
         CHAIN_CONNECT = createModel("chain_connect", "_connect", TextureKey.ALL);
-        TEMPLATE_SMALL_HEDGE_END = createModel("template_small_hedge_end", "_end", BlockusTextureKey.HEDGE);
+        TEMPLATE_SMALL_HEDGE_NOSIDE = createModel("template_small_hedge_noside", "_noside", BlockusTextureKey.HEDGE);
+        TEMPLATE_SMALL_HEDGE_NOSIDE_ALT = createModel("template_small_hedge_noside_alt", "_noside_alt", BlockusTextureKey.HEDGE);
+        TEMPLATE_SMALL_HEDGE_POST = createModel("template_small_hedge_post", "_post", BlockusTextureKey.HEDGE);
         TEMPLATE_SMALL_HEDGE_SIDE = createModel("template_small_hedge_side", "_side", BlockusTextureKey.HEDGE);
-        TEMPLATE_SMALL_HEDGE_SIDE_TALL = createModel("template_small_hedge_side_tall", "_side_tall", BlockusTextureKey.HEDGE);
+        TEMPLATE_SMALL_HEDGE_SIDE_ALT = createModel("template_small_hedge_side_alt", "_side_alt", BlockusTextureKey.HEDGE);
         TEMPLATE_SMALL_HEDGE_INVENTORY = createModel("template_small_hedge_inventory", "_inventory", BlockusTextureKey.HEDGE);
         CRATE_TEMPLATE = createModel("crate_template", TextureKey.TOP, TextureKey.SIDE);
         NEON_BLOCK_TEMPLATE = createModel("neon_block_template", TextureKey.ALL);