+++ changelog.md
@@ -1,1 +1,3 @@
 * Updated to 1.21.6
+* Improved small hedge block and model to make them more visually consistent (Small hedges already placed will need to be updated)
+* Small Hedges can now connect to leaves